Function minerva_rpc::session::make_client

pub async fn make_client(
    endpoint: String,
    tenant: String,
    requestor: String
) -> Result<SessionInterceptedClient, Status>

Creates a Session client with tenant and requestor metadata and connects to the server. Upon failure, returns an Unavailable gRPC status.

This function requires information about tenant, requestor and the actual endpoint that the client should connect to.