Module minerva_rpc::messages

Structures and algorithms related to all messages that can be sent to and from gRPC services and clients.

Structs

EntityIndex

A message containing the index of an expected entity.

PageIndex

A message containing the index of a page for a list of entities.

Product

A message describing a product that can be sent back and forth. This message represents products for inclusion, updating and retrieval.

ProductList

A message containing a list of products.

SessionCreationData

A message containing data to require a session creation on the backend.

SessionData

A message containing session data retrieved from the backend.

SessionToken

A message for returning a session token for the user.

User

A message describing a user that can be sent back and forth. This message represents users for inclusion, updating and retrieval.

UserList

A message containing a list of users of any size.